+++ src/main/java/org/vesalainen/grammar/impl/SyntheticParser.java
